+++ src/main/java/com/woowa/woowakit/domain/stock/application/StockService.java
@@ -5,11 +5,11 @@ import java.time.LocalDate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.woowa.woowakit.domain.model.ExpiryDate;
 import com.woowa.woowakit.domain.model.Quantity;
 import com.woowa.woowakit.domain.product.domain.Product;
 import com.woowa.woowakit.domain.product.domain.ProductRepository;
 import com.woowa.woowakit.domain.product.exception.ProductNotExistException;
-import com.woowa.woowakit.domain.stock.domain.ExpiryDate;
 import com.woowa.woowakit.domain.stock.domain.Stock;
 import com.woowa.woowakit.domain.stock.domain.StockRepository;
 import com.woowa.woowakit.domain.stock.dto.request.StockCreateRequest;
