+++ src/main/java/com/woowa/woowakit/domain/stock/domain/StockRepository.java
@@ -9,6 +9,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import com.woowa.woowakit.domain.model.ExpiryDate;
 import com.woowa.woowakit.domain.model.Quantity;
 
 public interface StockRepository extends JpaRepository<Stock, Long> {
