+++ src/test/java/com/woowa/woowakit/domain/model/ExpiryDateTest.java
@@ -1,4 +1,4 @@
-package com.woowa.woowakit.domain.stock.domain;
+package com.woowa.woowakit.domain.model;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import com.woowa.woowakit.domain.stock.exception.StockExpiredException;
+import com.woowa.woowakit.domain.model.exception.ExpiredException;
 
 @DisplayName("ExpiryDate 단위 테스트")
 class ExpiryDateTest {
@@ -23,7 +23,7 @@ class ExpiryDateTest {
 	@DisplayName("당일 이전의 재고를 생성하면 예외가 발생한다.")
 	void createExceptionIfPast() {
 		assertThatCode(() -> ExpiryDate.from(LocalDate.now().minusDays(1)))
-			.isInstanceOf(StockExpiredException.class)
-			.hasMessage("소비 기한이 지난 재고 항목입니다.");
+			.isInstanceOf(ExpiredException.class)
+			.hasMessage("만료일이 지났습니다.");
 	}
 }
