+++ src/test/java/com/woowa/woowakit/domain/stock/application/StockProcessingServiceTest.java
@@ -13,10 +13,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import com.woowa.woowakit.domain.fixture.ProductFixture;
+import com.woowa.woowakit.domain.model.ExpiryDate;
 import com.woowa.woowakit.domain.model.Quantity;
 import com.woowa.woowakit.domain.product.domain.Product;
 import com.woowa.woowakit.domain.product.domain.ProductRepository;
-import com.woowa.woowakit.domain.stock.domain.ExpiryDate;
 import com.woowa.woowakit.domain.stock.domain.Stock;
 import com.woowa.woowakit.domain.stock.domain.StockRepository;
 import com.woowa.woowakit.domain.stock.domain.StockType;
