+++ src/test/java/com/woowa/woowakit/domain/stock/domain/StockRepositoryTest.java
@@ -10,6 +10,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
 import com.woowa.woowakit.domain.fixture.ProductFixture;
+import com.woowa.woowakit.domain.model.ExpiryDate;
 import com.woowa.woowakit.domain.product.domain.Product;
 import com.woowa.woowakit.domain.product.domain.ProductRepository;
 import com.woowa.woowakit.global.config.JpaConfig;
